Refactor, fix queries
