+++ pkg/transport/manager.go
@@ -9,8 +9,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/SkycoinProject/dmsg/netutil"
-
 	"github.com/SkycoinProject/dmsg/cipher"
 	"github.com/SkycoinProject/skycoin/src/util/logging"
 	"github.com/google/uuid"
@@ -130,6 +128,7 @@ func (tm *Manager) serveNetwork(ctx context.Context, netType string) {
 }
 
 func (tm *Manager) serve(ctx context.Context) {
+	// TODO(nkryuchkov): to get rid of this callback, we need to have method on future network interface like: `Ready() <-chan struct{}`
 	// some networks may not be ready yet, so we're setting a callback first
 	tm.n.OnNewNetworkType(func(netType string) {
 		tm.serveNetwork(ctx, netType)
@@ -176,39 +175,7 @@ func (tm *Manager) initTransports(ctx context.Context) {
 			tpID   = entry.Entry.ID
 		)
 		if _, err := tm.saveTransport(remote, tpType); err != nil {
-			if err != snet.ErrNetworkNotReady {
-				tm.Logger.Warnf("INIT: failed to init tp: type(%s) remote(%s) tpID(%s)", tpType, remote, tpID)
-			}
-
-			// in this case network of `tpType` is not ready which is not permanent, so we're
-			// setting up a goroutine which will try to establish this transport
-			// TODO (nkryuchkov): once transport is generalized we could set up some kind of channel to wait for the network to come up and try to reestablish transport after that. for now we're just retrying to establish transport which is not really efficient
-			tm.Logger.Warnf("INIT: failed to init tp: type(%s) remote(%s) tpID(%s), retrying...", tpType, remote, tpID)
-
-			go func(entry *EntryWithStatus) {
-				retrier := netutil.NewRetrier(tm.Logger, 1*time.Second, 10*time.Second, 0, 1)
-				err := retrier.Do(ctx, func() error {
-					var (
-						tpType = entry.Entry.Type
-						remote = entry.Entry.RemoteEdge(tm.Conf.PubKey)
-					)
-
-					if _, err := tm.saveTransport(remote, tpType); err != nil {
-						if err == snet.ErrNetworkNotReady {
-							// keep on retrying only if the network is not ready.
-							// otherwise transport is either established or considered dead
-							return err
-						}
-					} else {
-						tm.Logger.Infof("Successfully initialized TP %v", *entry.Entry)
-					}
-
-					return nil
-				})
-				if err != nil {
-					tm.Logger.WithError(err).Warnf("INIT: failed to init tp: type(%s) remote(%s) tpID(%s)", tpType, remote, tpID)
-				}
-			}(entry)
+			tm.Logger.Warnf("INIT: failed to init tp: type(%s) remote(%s) tpID(%s)", tpType, remote, tpID)
 		} else {
 			tm.Logger.Debugf("Successfully initialized TP %v", *entry.Entry)
 		}
@@ -320,10 +287,6 @@ var ()
 func (tm *Manager) saveTransport(remote cipher.PubKey, netName string) (*ManagedTransport, error) {
 	if !snet.IsKnownNetwork(netName) {
 		return nil, snet.ErrUnknownNetwork
-	}
-
-	if !tm.n.IsNetworkReady(netName) {
-		return nil, snet.ErrNetworkNotReady
 	}
 
 	tpID := tm.tpIDFromPK(remote, netName)
